Refactor: Move PluginRepository to be under ui/jetpack for now.
More details on this conversation: https://github.com/woocommerce/woocommerce-android/pull/5287#pullrequestreview-810997145
